refactor: establish homeboy-refactor-contract crate keystone
Creates homeboy-refactor-contract and seeds it with the pure autofix-result type
cluster: AppliedRefactor, FixResultsSummary, RuleFixCount, PrimitiveFixCount.
These are behavior-free serde types describing the outcome of applying refactor
fixes; they're carried and constructed by consumers like the extension lint/test
report layer, which shouldn't need to depend on the refactor engine just for the
result vocabulary.

- New crate deps only serde/serde_json (pure leaf, sits below core).
- homeboy-core re-exports the types (crate::refactor::AppliedRefactor,
  refactor::auto::outcome::{FixResultsSummary, ...}) so existing call sites are
  unchanged.
- AppliedRefactor::from_capture (inherent method, no external callers) becomes a
  free fn refactor::applied_refactor_from_capture, since the contract type can't
  carry an engine-internal impl (it consumes the refactor-internal
  AppliedAutofixCapture).

This is the keystone seam for draining the extension -> refactor type edge (5
extension lint/test files import crate::refactor::AppliedRefactor). Same
keystone-first discipline as audit-contract before code_audit and
extension-contract: build the seam below core, then repoint consumers over time.

Also completes a pre-existing broken cli test constructor: config.agent_task.
rotation became Option<serde_json::Value> on main but promotion_review.rs still
assigned a typed AgentTaskProviderRotationPolicy (wrapped in serde_json::to_value
now). Verified broken identically on clean origin/main, unrelated to this change.

Refs #8425
